requirements.txt
aioredis[hiredis]==2.0.1
asks==3.0.0
asyncclick==8.1.3.4
trio==0.22.2
python-dotenv==1.0.0
quart-trio==0.11.0
pydantic==2.4.2
